updating + dialog changes
git-svn-id: http://svn.airdcpp.net/airsvn/trunk@1052 ec3e54a6-43b8-4a3e-9c49-2a733a73d9fe

+++ Localization.cpp
@@ -161,7 +161,7 @@ namespace dcpp {
 	}
 
 	void Localization::setLanguage(int languageIndex) {
-		if (languageIndex >= 0 && languageIndex < languageList.size() && languageList[languageIndex].languageName != languageList[curLanguage].languageName) {
+		if (languageIndex >= 0 && languageIndex < (int)languageList.size() && languageList[languageIndex].languageName != languageList[curLanguage].languageName) {
 			curLanguage = languageIndex;
 			languageList[curLanguage].setLanguageFile(); //update the language file in the settings
 		}
